requirements.txt
streamlit
pymongo
streamlit-extras
pillow
streamlit-shadcn-ui
